requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord.py==2.3.2
docker==7.0.0
docker-pycreds==0.4.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
packaging==23.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.1.0
websocket-client==1.7.0
yarl==1.9.4
